Move Snap() function to new `snap` package
Rename to snap.Take().

This function imports a lot of dependencies (such as os and image/*). Most games will not use it, so the function should not be a part of core pi package.
