+++ print_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/elgopher/pi"
+	"github.com/elgopher/pi/snap"
 	"github.com/elgopher/pi/vm"
 )
 
@@ -141,7 +142,7 @@ func TestPrint(t *testing.T) {
 func assertScreenEqual(t *testing.T, file string) {
 	expected := decodePNG(t, file).Pixels
 	if !assert.Equal(t, expected, vm.ScreenData) {
-		screenshot, err := pi.Snap()
+		screenshot, err := snap.Take()
 		require.NoError(t, err)
 		fmt.Println("Screenshot taken", screenshot)
 	}
